error handling

+++ app/models/csv_reader.rb
@@ -2,11 +2,16 @@
 class CsvReader
 
 	def initialize(csv_text)
-		@csv = CSV.parse(csv_text, :headers => true)
+		begin
+			@csv = CSV.parse(csv_text, :headers => true)
+		rescue CSV::MalformedCSVError #check for malformed csv
+			@error = true
+		end
 	end
 
 	#CSV format is 2 columns (name, partner name)
 	def process_names
+		return false if @error
 		User.delete_all #resets participants when you upload csv
 		0.upto(@csv.count-1) do |row|
 			user_name = @csv[row][0].presence
@@ -19,5 +24,7 @@ class CsvReader
 
 			User.find_or_create_by(name: partner_name.strip, partner: user) #create partner
 		end
+
+		return true
 	end
 end
